Resend data back to channel if no process is found
We're parsing raw syscall data on different goroutines.
We're getting process informations from execve event, but it may be
scheduled later, so we're not actually getting process info at that moment.

To fix this, `querySyscallEventToOPA` is run as seperate goroutines, so
we after parsing raw syscall data, we can immediately pass that syscall
event data to `opaQueryCh` channel and begin processing other data.
After receiving data from that channel, it searches looks for process
data in processMap, if it is not present, it sleeps for 100ms to wait for
process data become available and passes to `opaQueryCh` again